Revert "Session 3"
