+++ AdvancedProgrammingForTesters.sln
@@ -33,7 +33,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "DIP", "DIP", "{C7089D3B-52F
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ISP", "ISP\ISP.csproj", "{BD545324-8820-4CF8-A183-8C0923C7763C}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Test", "Test\Test.csproj", "{F051C2F9-5E18-43FA-8CB5-E3AEE75FF931}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DIP", "DIP\DIP.csproj", "{4DCA4A15-3796-472C-B9C6-C6A58213ED7C}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -77,10 +77,10 @@ Global
 		{BD545324-8820-4CF8-A183-8C0923C7763C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{BD545324-8820-4CF8-A183-8C0923C7763C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{BD545324-8820-4CF8-A183-8C0923C7763C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{F051C2F9-5E18-43FA-8CB5-E3AEE75FF931}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F051C2F9-5E18-43FA-8CB5-E3AEE75FF931}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F051C2F9-5E18-43FA-8CB5-E3AEE75FF931}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F051C2F9-5E18-43FA-8CB5-E3AEE75FF931}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4DCA4A15-3796-472C-B9C6-C6A58213ED7C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4DCA4A15-3796-472C-B9C6-C6A58213ED7C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4DCA4A15-3796-472C-B9C6-C6A58213ED7C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4DCA4A15-3796-472C-B9C6-C6A58213ED7C}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -98,7 +98,7 @@ Global
 		{2C2387ED-7E0C-48DE-A565-98D9CBFF8CF0} = {02EA681E-C7D8-13C7-8484-4AC65E1B71E8}
 		{C7089D3B-52F5-4A7A-BC6E-04ADBF2353C6} = {02EA681E-C7D8-13C7-8484-4AC65E1B71E8}
 		{BD545324-8820-4CF8-A183-8C0923C7763C} = {2C2387ED-7E0C-48DE-A565-98D9CBFF8CF0}
-		{F051C2F9-5E18-43FA-8CB5-E3AEE75FF931} = {C7089D3B-52F5-4A7A-BC6E-04ADBF2353C6}
+		{4DCA4A15-3796-472C-B9C6-C6A58213ED7C} = {C7089D3B-52F5-4A7A-BC6E-04ADBF2353C6}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {F6982061-311C-408C-B5CC-216002995EA4}
