--- AnimalType/ISoundMaker.cs
@@ -1,7 +1,0 @@
-﻿namespace AnimalType
-{
-    public interface ISoundMaker
-    {
-        public string MakeSound();
-    }
-}
