--- DIP/IButtonClicker.cs
@@ -1,8 +1,0 @@
-﻿namespace DIP
-{
-    public interface IButtonClicker
-    {
-        Task AcceptCookiesAsync();
-        Task RejectCookiesAsync();
-    }
-}
