+++ ISP/Interfaces/IDialogPopup.cs
@@ -1,7 +1,7 @@
-﻿using ISP.Interfaces;
-
-public interface IDialogPopup : IPopup
-{ 
+﻿public interface IDialogPopup
+{
+    void Show();
+    void Close();
     void SetTitle(string title);
     void SetMessage(string message);
 }
