--- ISP/Interfaces/IPopup.cs
@@ -1,8 +1,0 @@
-﻿namespace ISP.Interfaces
-{
-    public interface IPopup
-    {
-        void Show();
-        void Close();
-    }
-}
