+++ ISP/UploadPopup.cs
@@ -12,7 +12,21 @@
     {
         Console.WriteLine("Closing upload popup");
     }
- 
+
+    //from IDialogPopup
+    public void SetTitle(string title)
+    {
+        _title = title;
+        Console.WriteLine($"Title set: {title}");
+    }
+
+    //from IDialogPopup
+    public void SetMessage(string message)
+    {
+        _message = message;
+        Console.WriteLine($"Message set: {message}");
+    }
+
     public void SetUploadProgress(int percentage)
     {
         Console.WriteLine($"Upload progress: {percentage}%");
