--- Test/CookieTest.cs
@@ -1,20 +1,0 @@
-﻿using Microsoft.Playwright;
-
-namespace Test
-{
-    public class CookieTest
-    {
-        private readonly IButtonClicker _buttonClicker;
-
-        public CookieTest(IPage page)
-        {
-            _buttonClicker = new CookiePopup(page);
-        }
-
-        [Fact]
-        public async Task TestAcceptCookies()
-        {
-            await _buttonClicker.AcceptCookiesAsync();
-        }
-    }
-}
