--- Test/IButtonClicker.cs
@@ -1,8 +1,0 @@
-﻿namespace Test
-{
-    public interface IButtonClicker
-    {
-        Task AcceptCookiesAsync();
-        Task RejectCookiesAsync();
-    }
-}
